+++ crates/ruma-events/src/room/name.rs
@@ -1,7 +1,7 @@
 //! Types for the *m.room.name* event.
 
 use ruma_events_macros::EventContent;
-use ruma_identifiers::{RoomName, RoomNameBox};
+use ruma_identifiers::RoomNameBox;
 use serde::{Deserialize, Serialize};
 
 use crate::StateEvent;
@@ -23,12 +23,6 @@ impl NameEventContent {
     /// Create a new `NameEventContent` with the given name.
     pub fn new(name: Option<RoomNameBox>) -> Self {
         Self { name }
-    }
-
-    /// The name of the room, if any.
-    #[deprecated = "You can access the name field directly."]
-    pub fn name(&self) -> Option<&RoomName> {
-        self.name.as_deref()
     }
 }